Rename `DropDownTokenFilter` to consisent casing `DropdownTokenFilter`
